Updated the code to support negative prices and negative (decremental) energy usage
Added a calibration method which recalculates the cost at price change events (each hour)
The most precise method of cost calculation is use the delta energy for the past hour and multiple this with the electricity price of that hour.
Therefore at each price change event we calulate the cost for the past hour.
However, as this will only update the cost each per hour I kept the original code for updates in the running hour, and overwrite this value with a more precise value